requirements.txt
fastapi
uvicorn
twilio
python-dotenv
google-cloud-aiplatform
google-generativeai
soundfile
